ktlo(common): Inline Evm-Compat Impls Into Types (#2218)
* ktlo(common): inline evm-compat impls into types

* fix(common): fix nightly fmt import ordering for cfg-gated evm imports

* fix(common): remove blank line from lib.rs after rebase

* fix(builder-core): add backticks around MeteringDataPending in doc comment

* fix(builder-metering): add backticks around MeteringDataPending in doc comment


---------
